Relocate options-cache helper to plugin options
